Move createNonce to shared
