+++ packages/shared/src/utils/Error.ts
@@ -1,0 +1,6 @@
+export type ErrorCode = string;
+export type Web3Error = Error & { code: ErrorCode; customMessage?: string };
+
+export function isWeb3Error(error: unknown): error is Web3Error {
+  return typeof error === 'object' && error !== null && 'code' in error;
+}
